refactored the code for vlc on esc video play is not working
